refactor: move StorageSettings
