+++ src/scripts/dictionary/Dictionary.ts
@@ -2,7 +2,7 @@ import DictionaryView from "./DictionaryView";
 import DictionaryModel from "./DictionaryModel";
 import {WordsSettings} from "../sprint/SprintSettings";
 import Sprint from "../sprint/Sprint";
-import {LocalStorage} from "../../state/StrageSettings";
+import {LocalStorage} from "../state/StorageSettings";
 
 class Dictionary {
   private readonly model: DictionaryModel;
