+++ src/scripts/state/StorageSettings.ts
@@ -2,3 +2,7 @@ export enum LocalStorage {
   dictionaryDifficultyLevel = "dictionaryDifficultyLevel",
   dictionaryPageNumber = "dictionaryPageNumber",
 }
+
+export enum SessionStorage {
+  userData = "userData",
+}
